- Add support _framework_ for the Asus p2b. - New superIO winbond/w83977tf - Add single memory controller SBbus debug routine into a file private to the i440bx
This adds support the start of support for an Asus p2b
mainboard.  Current limitations are the same as for the 
Bitworks IMS board.  Reads from the SMbus don't work.

Moving dump_spd_registers() into its own private copy
solves the problem of having to go hack on the version that
included in src/sdram to only do one memory controller.




git-svn-id: svn://svn.coreboot.org/coreboot/trunk@2351 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1