Stupid malloc allocater which we hope that we will never need. But it has an author.
The use of malloc is very limited in V2, and we are pretty sure we can
reduce usage of malloc to no use at all in V3. 

The device code uses malloc, but that needs to be changed, so that the
device tree is allocated in a known-good-place in ram and can be
passed to other stages and payloads which may want to parse it. 

The built-in-filo and vfs code in v2 both use malloc, but they will not be
brought forward to V3. 

The old elfboot code used malloc, but we have changed that code so much
it no longer needs malloc. 

The only remaining use of malloc is lzma, and we see no need to use such 
a general mechanism for the lzma's single call to malloc. 

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Acked-by: Ronald G. Minnich <rminnich@gmail.com>
Acked-by: Stefan Reinauer <stepan@coresystems.de>



git-svn-id: svn://coreboot.org/repository/LinuxBIOSv3@113 f3766cd6-281f-0410-b1cd-43a5c92072e9