Initial cleanup of the linuxbios tables code. Now it seems we need resource .c functions -- sigh -- this will be next.
fix types, fix some usage, but we do not yet generate pirqi, mptables,
or acpi. This will be next. 

This code will be called from phase6 in stage 2, and probably 
attached to a cpu. Or, it will be inline in the stage2.c code, since 
the need for tables is generic. 

I'm also going to move the cryptic usage of lgdt from here to stage 2,
phase 2, of the cpu code. The device model improvements are making it
easier to think about when things happen, we are finding. 

I'm going to get lunch and go see some of brussels. It's a miserable day
but ... 

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Acked-by: Ronald G. Minnich <rminnich@gmail.com>
Acked-by: Stefan Reinauer <stepan@coresystems.de>




git-svn-id: svn://coreboot.org/repository/LinuxBIOSv3@130 f3766cd6-281f-0410-b1cd-43a5c92072e9