Add support for generating Doxygen-documentation via 'make doc'.
The output is in HTML format and will be stored in the doxygen/html
directory. You can enable more output formats (latex, RTF, ...)
in the config file, util/Doxyfile.LinuxBIOS.

The current config options seem pretty reasonable to me, but we can
adapt them further to our needs, if required.

Signed-off-by: Uwe Hermann <uwe@hermann-uwe.de>
Acked-by: Stefan Reinauer <stepan@coresystems.de>



git-svn-id: svn://coreboot.org/repository/LinuxBIOSv3@171 f3766cd6-281f-0410-b1cd-43a5c92072e9