This patch drives the idea of a hierarchical structure below $(obj) a bit further. It is only tested with my Kconfig build dir fixes applied.
Not sure how pretty this is. All the mkdirs increase build time by 0.2s,
so it looks a lot more nasty than it actually is. If wishes, we can create the
whole directory structure below build in the "prepare" target. It's not faster,
but more limited to one place.

The next step could be to spread out *_LIB_OBJ to lib/Makefile. in which case
it is nice that you can just write object.o instead of $(obj)/object.o or
$(obj)/lib/object.o ...

The top level modules (initram, stage0, stage2, ..) are still in $(obj).

So are the northbridge and southbridge object files - These require some more
cleanup anyways, as they're defined in the mainboard Makefile.

Signed-off-by: Stefan Reinauer <stepan@coresystems.de>
Acked-by: Uwe Hermann <uwe@hermann-uwe.de>



git-svn-id: svn://coreboot.org/repository/LinuxBIOSv3@242 f3766cd6-281f-0410-b1cd-43a5c92072e9