This fixes ram.c correctly for cases where we link instead of including .c files.
The mem controller is an opaque type. This code is called with a
pointer and a count of mem controllers.

I have serious doubts about the value of this code. It is so generic,
and does so little, that it
may be useless. It may be useful for documentation, however, as it
shows people the sequence of operations for spd.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Acked-by: Stefan Reinauer <stepan@coresystems.de>



git-svn-id: svn://coreboot.org/repository/LinuxBIOSv3@339 f3766cd6-281f-0410-b1cd-43a5c92072e9