This is a subtle error. An operations struct, to work as a constructor, must initialize the .constructor struct member.
This error is so subtle that we might consider changing the device code; or at least 
working on the documentation to make it more clear. The key question: do we ever want 
an operations struct that can NOT operate as a device constructor? When this code was 
designed over a year ago, it seemed that we might; now, it is not so certain. 

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>



git-svn-id: svn://coreboot.org/repository/coreboot-v3@577 f3766cd6-281f-0410-b1cd-43a5c92072e9