This is an emergency commit because the zero compression, as it stands, is causing coreboot on the alix1c to hang. I don't know why.
But this patch is the difference between works and hangs. I think 
the memset in the decompress is not using the right start value? Not sure. 

This simple change allows us to fix the zero decompress later, when 
we figure it out. I just lost a few hours to this and I want to make
sure the repo works before I go to bed. 

This fix tested on alix1c. Alix1c now works fine.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Acked-by: Ronald G. Minnich <rminnich@gmail.com>




git-svn-id: svn://coreboot.org/repository/coreboot-v3@604 f3766cd6-281f-0410-b1cd-43a5c92072e9