Bring 8111 over to v3.
Rename files and functions as needed. 

There is regularity to the naming. Stage1 stuff is called stage1_*. The rest is not. 
Most .c files have a corresponding .dts. The code is simpler and smaller and has less
duplication. 

Most (all) romcc artifacts removed. 

I've made a lot of effort to get copyright headers done correctly, using 'svn log'. 

next are the 8132, 8151, and serengeti, then comes simnow. 
 
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Acked-by: Ronald G. Minnich <rminnich@gmail.com>


git-svn-id: svn://coreboot.org/repository/coreboot-v3@767 f3766cd6-281f-0410-b1cd-43a5c92072e9